requirements.txt
pydantic>=2.7.0
beautifulsoup4>=4.12.0
langchain>=0.1.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
agentql
google-genai
browserbase
tiktoken
supabase
openai>=1.0.0
thefuzz
pytesseract
pillow
opencv-python
numpy
playwright
fastapi
uvicorn
python-multipart
python-dotenv